Refactor some elements
